Coasting & automatic non signal pathed trains
Catch coasting trains AND trains that path in to a portal's tracks, but not in to the tunnel itself. Check that they are allowed to enter the tunnel and reserve it, although reservation is lost as their speed doesn't reach the END signal. This does ensure the tunnel is available and they close the other entrance signals automatically.