enter train length ignores end connection distance
deleted force-repath-back-through-tunnel-tests as it was designed to handle the complexity of the old mod with trains being half in/out and trying to reverse with loco's still underground etc. But new logic's simple reversing option is tested as part of remove-target-stop-rails test.
remaining un-used tests are for bugs explicitly listed in TODO.