Tunnel approach downgrade to onPortalTrack - WIP
Reduced all tests that set train speed to 2 down to 1.4. As 2 would trigger the tunnel and then a tick later the trains speed reduces and the tunnel was aborted. 1.4 is a locomotive max.
Speed duration test train starting distances changed to avoid false tunnel triggers on first tick for fast trains.