Test settings ease of use
Make tests behave so that DoSpecificTests overwrites DoMinimalTests. As DoMinimalTests is on by default on tests and DoSpecificTests is off by default. When wanting to do an adhoc test its easier to just turn on 1 setting than have to also turn one off.
The weird issue with reversing train having an extra temporary stop was caused by the previously resolved bug. Tested and no longer occurs.