Player riding train - WIP 9
Removed the caching of breaking point as while it did save UPS it could have lead to odd situations where the train nearly overshoots the end of the tunnel if the end portal rail signal suddenly changes to closed at the wrong time. 
Few users will be riding a train at a time and the actual UPS cost per user is very low.