Filter left clicks on portal if something in cursor
When using a BP or copy/cut and you start by clicking on a tunnel portal part it did open the portal GUI. Now these are ignored as well as if the player has anything in their cursor. This is the same as if a player has a cursor item and clicks on any other openable entity.