Fix mine tunnel part leaving train destroyed edge case
When trying to mine a portal or underground part continously (hold right click in editor) as a train is using the tunnel, sometimes as it leaves the portal the tunnel part will allow the mine and still destroy the train. This is due to where the portal entry detector has to be placed and we need to check if the entry end rail of each portal can be destroyed (no train on them).